+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/AbstractFastTigerProxyTest.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright (c) 2023 gematik GmbH
+ * 
+ * Licensed under the Apache License, Version 2.0 (the License);
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an 'AS IS' BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package de.gematik.test.tiger.proxy;
+
+import de.gematik.test.tiger.common.data.config.tigerProxy.TigerProxyConfiguration;
+import de.gematik.test.tiger.config.ResetTigerConfiguration;
+import lombok.extern.slf4j.Slf4j;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
+
+@Slf4j
+@TestInstance(Lifecycle.PER_CLASS)
+@ResetTigerConfiguration
+class AbstractFastTigerProxyTest extends AbstractTigerProxyTest {
+
+    @BeforeAll
+    public void setupTigerProxy() {
+        spawnTigerProxyWith(TigerProxyConfiguration.builder()
+            .build());
+    }
+
+    @BeforeEach
+    public void cleanupTiger() {
+        tigerProxy.clearAllRoutes();
+        tigerProxy.clearAllMessages();
+    }
+
+    @Override
+    public void stopSpawnedTigerProxy() {
+        // nothing
+    }
+}